python-consumer/requirements.txt
pika==1.2.0
pymongo[srv]
python-dotenv==0.19.2
